// web/src/types.ts
export interface ModHit {
  project_id: string
  slug: string
  title: string
  description: string
  author: string
  downloads: number
  follows: number
  icon_url: string | null
  categories: string[]
  versions: string[]
  client_side: string
  server_side: string
  project_type: string
}